Move the shelf into its own file and drop the comments
